Reorganized new Hierarchy API
